renamed money module to economy
